requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
python-multipart==0.0.6
numpy==1.21.0
Pillow==10.0.1
anthropic==0.30.0
python-dotenv==1.0.0
deepface==0.0.79